_io: take the seek whence as a C int on the index protocol
`BytesIO.seek` and `StringIO.seek` unwrapped their whence with the machine-int
converter, so a value outside a C int's range reached the `0/1/2` range check
and came back as `ValueError: invalid whence (4294967296, should be 0, 1 or 2)`.

Both `@unwrap_spec(pos=r_longlong, whence=int)` (interp_bytesio.py:162) and
`@unwrap_spec(pos=int, mode=int)` (interp_stringio.py:403) take it as a C int,
and 3.14 raises `OverflowError: Python int too large to convert to C int` for
one that does not fit.

Neither existing converter is that parameter. `c_int_w` has the range check but
reaches the value through `gateway_int_w` = `int_w` (baseobjspace.py:2043),
which converts through `__int__` first; `index_int_w_preserve_negative` is on
the index protocol but is a machine int. A whence needs both, so add
`index_c_int_w` — `space_index` followed by the 32-bit check — and the
`PyIndexCInt` unwrap alias over it, and use it at both sites.

Reading the whence through `__int__` is observable, not merely a message
difference: an object defining `__index__` as 0 and `__int__` as 2 seeks to the
end rather than the start, silently, and one defining only `__int__` is accepted
where 3.14 raises TypeError. The parity script asserts the whence by value
against a non-empty buffer so neither case can pass, and pins the out-of-range
half and the position's wider `Py_ssize_t` range beside it.

The position is left alone: upstream's `pos=int` on `StringIO` is a C int too,
but 3.14 takes it as a `Py_ssize_t` and accepts `seek(2**32)`.

+++ pyre/pyre-interpreter/src/module/_io/bytesio.rs
@@ -389,9 +389,14 @@ impl W_BytesIO {
     fn seek(
         &mut self,
         pos: PyIndexInt,
-        #[default(0)] whence: PyIndexInt,
+        #[default(0)] whence: PyIndexCInt,
     ) -> Result<i64, crate::PyError> {
         // interp_bytesio.py:162-180 validation followed by RStringIO.seek.
+        // `@unwrap_spec(pos=r_longlong, whence=int)` (:162): the position is a
+        // long long but the whence is a C int, so one that does not fit is an
+        // OverflowError from the converter rather than a value the arms below
+        // ever see. Both come through `__index__` alone — 3.14 rejects a whence
+        // carrying only `__int__`, and reads one carrying both from `__index__`.
         self.check_closed()?;
         match whence {
             0 if pos < 0 => {
@@ -416,7 +421,7 @@ impl W_BytesIO {
                 )));
             }
         }
-        self.seek_pos(pos, whence);
+        self.seek_pos(pos, whence as i64);
         Ok(self.tell_pos())
     }
